1.商品的分类  2.商品随机id的添加

+++ manager-pojo/src/main/java/com/shengde/e3mall/entity/TbItemCat.java
@@ -14,7 +14,8 @@ public class TbItemCat implements Serializable {
 
     private Integer id;
 
-    private Integer parentId;
+
+    private long parentId;
 
     private String name;
 
@@ -37,13 +38,14 @@ public class TbItemCat implements Serializable {
         this.id = id;
     }
 
-    public Integer getParentId() {
+    public long getParentId() {
         return parentId;
     }
 
-    public void setParentId(Integer parentId) {
+    public void setParentId(long parentId) {
         this.parentId = parentId;
     }
+
 
     public String getName() {
         return name;
